bugfix for resp parser (#4)
Fixes #3 by improving the utlity functions which we use to operate
on the array of tokens.

Previously, we were having to track our position in the token
stream but were mixing that with functions which removed tokens
from the stream.

This change makes all the utility functions act by attempting to
take a token from the front of the token stream. This makes the
parser more intuitive to write and less likely to have the same
class of bug.