chore: Update CI to improve throughput and caching (#61)
This PR is meant to improve the throughput of CI and reduce its cache footprint. This way, if multiple PRs are being updated at the same time then they'll spend less time waiting. It does not reduce the E2E latency of the cache pipline, instead it bumps jobs that are less important to the end and focuses on reducing total compute time.

Detailed changes:
- Avoid using any actions from `actions-rs` because they have not been maintained in years
- Move all builds for fuzz targets to a single job and cache entry (fuzzing still runs in separate jobs)
- Share caches between jobs in the regular cargo workflow where possible
- Capture clippy annotations and expose them via the GH UI using `clippy-sarif`
- Smoketest jobs no longer build their binaries, instead they use binaries compiled by the ubuntu-latest `build` job
- I've removed `bench-check` since it does nothing that isn't covered by `build` or `check`