chore: add .dockerignore to speed up local builds
This shouldn't improve anything in a CI-context i.e. any time you're
cloning the repository fresh and running docker commands inside.

However, with local development, we are currently sending all of
`target` in the Docker "build context" which, on my machine is 18GB and
dramatically slows down the Docker build loop.