fix: save ourselves one syscall per request on unix
Since we use level-triggered events, we previously were calling
read() until we got EWOULDBLOCK. This strictly follows the docs for
mio.

However, on unix systems, we can rely on the fact that if we read
less data than we wanted, that we will get a EWOULDBLOCK on the
next read.

This change adds a block that will early return in this case for
"unix" flavored targets and results in saving one read() per
request.