feat: Introduce new RESP parser (#64)
This PR introduces a rewrite of the resp parser with a couple of goals:
- Avoid having to build an intermediate `Message` structure when we can just parse the messages directly
- Simplify the boilerplate around parsing each RESP command
- And finally, be a bit more understandable if someone needs to debug/modify the code

This commit introduces two main types under the new `parse` module
- `Parser` is a complete parser for the resp protocol
- `CommandParser` is a stub parser that parses both the RESP command format as well as the inline format

I have modeled the API off of a combination of syn and serde. It makes no allocations whatsoever in the parsing phase while leaving the option to continue parsing to `Message` if that is needed.

The one bit that is a bit of a mess is the `ParseError` type. I will be coming back to clean this up in a future PR when I get the time. As is, it is good enough for now, so it shouldn't be much of an issue.