feat: prepare seg storage crate for release
Prepare the seg storage crate for release as `pelikan-storage-seg`.

Changes to make the metrics an optional, but default feature to
help reduce the dependency tree for embedded cases.

Remove the dependency on the `storage-types` by adding the typed
`Value`s into this crate itself.

Supporting changes for rename and versioning.