deps: update histogram, metriken, and ringlog (#129)
Update dependencies to move towards a single version
of histogram (0.11.0) and metriken (0.7.0). This is
a breaking change in the interface of calls to percentile
functions, which can now return null when the histogram
is empty.

The metrics functions call the percentiles and publish
them using an iterator over them. Rather than changing
the interface of the wrapper function to returning null
when the histogram is empty, it returns an empty vector
to metrics callers, which prevents them from publishing
percentile metrics.