fix: translate get `NotFound` to `END` for text protocol
In a recent refactor the memcached binary protocol was introduced. In
the binary protocol, `NOT_FOUND` is relevant for `GET` misses, unlike
for the text protocol. While the refactor correctly implemented the
binary protocol case, it introduced a bug in the text protocol case.

For the text protocol `GET` where all keys `MISS`, the result should
be a plain `END\r\n` as if for an empty set of values. Instead it was
returning `NOT_FOUND`, which is not to spec.